backend/requirements.txt
fastapi
uvicorn[standard]
proxmoxer
requests
python-dotenv